change labelName to name + add label check for wrapped name
